try to improve diffs
